agrega los objetos al array takeOrdern
