+++ burger-queen-project/src/pages/breakfast.jsx
@@ -1,6 +1,5 @@
-import { HeaderMenu } from "../components/menu/headerMenu";
+import { HeaderMenu } from '../components/menu/headerMenu' 
 import { NavigationMenu } from "../components/menu/navigationMenu";
-// import { BoxCategoriesMenuFood } from "../components/menu/boxCategoriesMenuFood";
 import { ViewAllFood } from "../components/menu/viewAllFood";
 import { OrderProvider } from "../components/menu/contextOrder";
