+++ burger-queen-project/src/pages/food.jsx
@@ -1,4 +1,4 @@
-import { HeaderMenu } from "../components/menu/headerMenu";
+import { HeaderMenu } from '../components/menu/headerMenu' 
 import { NavigationMenu } from "../components/menu/navigationMenu";
 // import { BoxCategoriesMenuFood } from "../components/menu/boxCategoriesMenuFood";
 import { ViewAllFood } from "../components/menu/viewAllFood";
