+++ burger-queen-project/src/pages/menu.jsx
@@ -1,16 +1,17 @@
-import { HeaderMenu } from "../components/menu/headerMenu";
+import { HeaderMenu } from '../components/menu/headerMenu' 
 import { NavigationMenu } from "../components/menu/navigationMenu";
 import { BoXBurgerDance } from "../components/menu/boxBurgerDance";
 import { BoxOrder } from "../components/menu/boxOrder";
 import { BtnProcessOrder } from "../components/menu/btnProcessOrder";
-
-// import { AllFood } from "../components/menu/all-food";
+import { OrderProvider } from "@/components/menu/contextOrder";
 
 export default function MenuPage() {
     return (<div>
         <NavigationMenu />
         <BoXBurgerDance />
+        <OrderProvider>
         <BoxOrder />
+        </OrderProvider>
         <BtnProcessOrder />
     </div>
     )
